Move _read_events() to read.py
