+++ mne_bids/read.py
@@ -12,7 +12,7 @@ import json
 
 import numpy as np
 import mne
-from mne import io
+from mne import io, read_events, events_from_annotations
 from mne.utils import has_nibabel, logger, warn
 from mne.coreg import fit_matched_points
 from mne.transforms import apply_trans
@@ -61,6 +61,77 @@ def _read_raw(raw_fpath, electrode=None, hsp=None, hpi=None,
                          f'of {ALLOWED_DATATYPE_EXTENSIONS}\n'
                          f'Got {ext}')
     return raw
+
+
+def _read_events(events_data, event_id, raw, verbose=None):
+    """Retrieve events (for use in *_events.tsv) from FIFF/array & Annoations.
+
+    Parameters
+    ----------
+    events_data : str | array | None
+        The events file. If a string, a path to the events file. If an array,
+        the MNE events array (shape n_events, 3). If None, events will be
+        inferred from the stim channel using `find_events`.
+    event_id : dict
+        The event id dict used to create a 'trial_type' column in events.tsv,
+        mapping a description key to an integer valued event code.
+    raw : instance of Raw
+        The data as MNE-Python Raw object.
+    verbose : bool | str | int | None
+        If not None, override default verbose level (see :func:`mne.verbose`).
+
+    Returns
+    -------
+    events : array, shape = (n_events, 3)
+        The first column contains the event time in samples and the third
+        column contains the event id. The second column is ignored for now but
+        typically contains the value of the trigger channel either immediately
+        before the event or immediately after.
+    descriptions : dict
+        A dictionary with the keys corresponding to the event descriptions and
+        the values to the event IDs.
+
+    """
+    # get events from events_data
+    if isinstance(events_data, str):
+        events = read_events(events_data, verbose=verbose).astype(int)
+    elif isinstance(events_data, np.ndarray):
+        if events_data.ndim != 2:
+            raise ValueError('Events must have two dimensions, '
+                             f'found {events_data.ndim}')
+        if events_data.shape[1] != 3:
+            raise ValueError('Events must have second dimension of length 3, '
+                             f'found {events_data.shape[1]}')
+        events = events_data
+    else:
+        events = np.empty(shape=(0, 3), dtype=int)
+
+    # Only keep events for which we have an ID <> description mapping.
+    # XXX Should we emit a warning if we drop events?
+    mask = [e in list(event_id.values()) for e in events[:, 2]]
+    events = events[mask]
+
+    # Append events to raw.annotations.
+    #
+    # Appending to the existing Annotations object frees us from the
+    # requirement to pass the `orig_time` parameter, as it's already been set
+    # correctly. All event onsets are relative to measurement beginning.
+    if events.size > 0:
+        id_to_desc_map = dict(zip(event_id.values(), event_id.keys()))
+        onsets = events[:, 0] / raw.info['sfreq']
+        durations = np.zeros_like(onsets)  # Instantaneous events.
+        descriptions = [id_to_desc_map[event_id] for event_id in events[:, 2]]
+        for onset, dur, desc in zip(onsets, durations, descriptions):
+            raw.annotations.append(onset=onset, duration=dur, description=desc)
+        del id_to_desc_map, onsets, durations, descriptions
+
+    # Now convert the Annotations to events.
+    all_events, all_desc = events_from_annotations(raw, verbose=verbose)
+    if all_events.size == 0:
+        warn('No events found or provided. Please add annotations to the raw '
+             'data, or provide the events_data and event_id parameters.')
+
+    return all_events, all_desc
 
 
 def _handle_participants_reading(participants_fname, raw,
